model stuff

+++ selenobot/datasets.py
@@ -57,7 +57,7 @@ class Dataset(torch.utils.data.Dataset):
             self.labels = torch.from_numpy(df['label'].values).type(torch.LongTensor)
             self.labels_one_hot_encoded = one_hot(self.labels, num_classes=n_classes).to(self.device)
 
-        self.embeddings = torch.from_numpy(df[Dataset.get_feature_cols(df)].values).to(self.device).to(self.dtype)
+        self.embeddings = torch.from_numpy(df[Dataset.get_feature_cols(df)].values).to(self.device)
         self.n_features = len(feature_cols)
         self.metadata = df[[col for col in df.columns if type(col) == str]] 
         self.ids = df.index.values
